week10/EducationSystem/requirements.txt
Django==3.1
psycopg2-binary
Pillow==7.1
djangorestframework==3.11

